backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]>=3.2.10
pydantic-settings==2.5.2
python-dotenv==1.0.1
alembic==1.13.3
pyjwt==2.9.0
bcrypt==4.2.0
